--- src/components/mypage/SummarizedCollection/index.ts
@@ -1,1 +1,0 @@
-export * from "./SummarizedCollection";
