+++ src/features/mypage/components/SummarizedCollection.tsx
